+++ mex/editor/merge/state.py
@@ -10,7 +10,7 @@ from mex.common.models import MERGED_MODEL_CLASSES
 from mex.common.transform import ensure_prefix
 from mex.editor.exceptions import escalate_error
 from mex.editor.label_var import label_var
-from mex.editor.search.models import SearchResult
+from mex.editor.models import SearchResult
 from mex.editor.search.transform import transform_models_to_results
 from mex.editor.state import State
 from mex.editor.utils import resolve_editor_value