+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Component to display search results
+
 ### Changes
+
+- switch package manager from pdm to uv
 
 ### Deprecated